+++ test/pay/stripe/billable_test.rb
@@ -24,4 +24,10 @@ class Pay::Stripe::BillableTest < ActiveSupport::TestCase
     assert_not_nil invoice.lines.data.find { |l| l.plan&.id == "default" }
     assert_not_nil invoice.lines.data.find { |l| l.price&.id == "price_1ILVZaKXBGcbgpbZQ26kgXWG" }
   end
+
+  test "stripe saves currency on charge" do
+    @user.card_token = "pm_card_visa"
+    charge = @user.charge(29_00)
+    assert_equal "usd", charge.currency
+  end
 end