+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Save `currency` on Pay::Charge records - @excid3
+
 ### 2.6.11
 
 * Add `subscription` method to payment processor classes for direct access to the processor subscription object.